Move useReplayFilter hook into lib/hooks
